Enum elmesque::color::Gradient

pub enum Gradient {
    Linear((f64, f64), (f64, f64), Vec<(f64, Color)>),
    Radial((f64, f64), f64, (f64, f64), f64, Vec<(f64, Color)>),
}

Linear or Radial Gradient.

Variants

Linear

Takes a start and end point and then a series of color stops that indicate how to interpolate between the start and end points.

Radial

First takes a start point and inner radius. Then takes an end point and outer radius. It then takes a series of color stops that indicate how to interpolate between the inner and outer circles.