Function elmesque::color::light_red

pub fn light_red() -> Color

Built-in colors.

These colors come from the Tango palette which provides aesthetically reasonable defaults for colors. Each color also comes with a light and dark version. Scarlet Red - Light - #EF2929