Function elmesque::form::draw_form

pub fn draw_form<'a, C: CharacterCache, G: Graphics<Texture=C::Texture>>(form: &Form, alpha: f32, backend: &mut G, maybe_character_cache: &mut Option<&mut C>, context: Context)

CUSTOM NON-ELM FUNCTIONS.

Normally Elm renders to html and javascript, however the aim of elmesque is to render to GL.

This function draws a form with some given transform using the generic Piston graphics backend.