Module elmesque::text

Structs

Style
Represents all the ways you can style Text. If the type_face list is empty or the height is None, the users will fall back on their default settings. The following Style is black, 16 pixel tall, underlined, and Times New Roman (assuming that typeface is available on the user's computer):

Text
Drawable Text.

TextUnit

Enums

Line
Styles for lines on text. This allows you to add an underline, an overline, or strike out text.

Position
Text position relative to center point